LayoutTests/imported/w3c/web-platform-tests/shadow-dom/declarative/tentative/shadowrootadoptedstylesheets/shadowrootadoptedstylesheets-async-fetch-failure-expected.txt


FAIL Async fetch failure: 404 leaves placeholder empty, valid specifier still works. assert_equals: Before fetch settles: expected 1 placeholder(s). expected 1 but got 0
